SortPoints and Perspektive Transform Moved from CardAanlyser to imageOperator.py.
